Moving reducers/actions/constants to hearingSchedule folder
