+++ client/app/hearingSchedule/actions/actions.js
@@ -1,4 +1,4 @@
-import { ACTIONS } from './constants';
+import { ACTIONS } from '../constants';
 
 export const onReceivePastUploads = (pastUploads) => ({
   type: ACTIONS.RECEIVE_PAST_UPLOADS,
