+++ client/app/hearingSchedule/index.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReduxBase from '@department-of-veterans-affairs/caseflow-frontend-toolkit/components/ReduxBase';
-import reducers, { initialState } from './reducers';
+import reducers, { initialState } from './reducers/reducers';
 import HearingsApp from './HearingsApp';
 
 const Hearings = (props) => <ReduxBase store={initialState} reducer={reducers}>
