+++ client/app/hearingSchedule/reducers/reducers.js
@@ -1,14 +1,14 @@
 /* eslint-disable max-lines */
-import { timeFunction } from '../util/PerfDebug';
-import { ACTIONS } from './constants';
-import { update } from '../util/ReducerUtil';
+import { timeFunction } from '../../util/PerfDebug';
+import { ACTIONS } from '../constants';
+import { update } from '../../util/ReducerUtil';
 import { combineReducers } from 'redux';
 
-import commonComponentsReducer from '../components/common/reducers';
-import caseListReducer from '../queue/CaseList/CaseListReducer';
-import { workQueueReducer } from '../queue/reducers';
-import hearingsReducer from '../hearings/reducers/index';
-import uiReducer from '../queue/uiReducer/uiReducer';
+import commonComponentsReducer from '../../components/common/reducers';
+import caseListReducer from '../../queue/CaseList/CaseListReducer';
+import { workQueueReducer } from '../../queue/reducers';
+import hearingsReducer from './index';
+import uiReducer from '../../queue/uiReducer/uiReducer';
 
 export const initialState = {};
 const hearingScheduleReducer = (state = initialState, action = {}) => {
