+++ client/app/hearings/components/HearingWorksheetStream.jsx
@@ -4,7 +4,7 @@ import { bindActionCreators } from 'redux';
 import _ from 'lodash';
 import Button from '../../components/Button';
 import TabWindow from '../../components/TabWindow';
-import { onAddIssue } from '../actions/Issue';
+import { onAddIssue } from '../../hearingSchedule/actions/Issue';
 import HearingWorksheetIssues from './HearingWorksheetIssues';
 
 class HearingWorksheetStream extends Component {
