+++ client/app/hearings/components/WorksheetHeader.jsx
@@ -6,10 +6,10 @@ import classNames from 'classnames';
 import Textarea from 'react-textarea-autosize';
 import { ClipboardIcon } from '../../components/RenderFunctions';
 import CopyToClipboard from 'react-copy-to-clipboard';
-import { onRepNameChange, onWitnessChange, onMilitaryServiceChange } from '../actions/Dockets';
+import { onRepNameChange, onWitnessChange, onMilitaryServiceChange } from '../../hearingSchedule/actions/Dockets';
 import { css } from 'glamor';
 import _ from 'lodash';
-import { DISPOSITION_OPTIONS } from '../constants/constants';
+import { DISPOSITION_OPTIONS } from '../../hearingSchedule/constants';
 import Tooltip from '../../components/Tooltip';
 import DocketTypeBadge from '../../components/DocketTypeBadge';
