+++ client/app/queue/components/AssignHearingModal.jsx
@@ -21,7 +21,7 @@ import {
 import { onReceiveAmaTasks, onReceiveAppealDetails } from '../QueueActions';
 import { prepareAppealForStore } from '../utils';
 import _ from 'lodash';
-import { CENTRAL_OFFICE_HEARING, VIDEO_HEARING } from '../../hearings/constants/constants';
+import { CENTRAL_OFFICE_HEARING, VIDEO_HEARING } from '../../hearingSchedule/constants';
 import QueueFlowModal from './QueueFlowModal';
 import AssignHearingForm from '../../hearingSchedule/components/modalForms/AssignHearingForm';
